// src/components/ui/paper-onboarding/types.d.ts
export interface PaperOnboardingItemType {
  /**
   * Image cover.
   * @type {(() => React.ReactNode) | React.ReactNode}
   */
  image?: (() => ReactNode) | ReactNode;
  /**
   * Indicator icon.
   * @type {((props: IndicatorIconProps) => React.ReactNode) | React.ReactNode}
   */
  icon?: FC<IndicatorIconProps> | ComponentClass<IndicatorIconProps>;
  /**
   * Slide/page content.
   * @type {((props: PageContentProps) => React.ReactNode) | React.ReactNode}
   */
  content?: FC<PageContentProps> | ComponentClass<PageContentProps>;
  /**
   * Title text.
   * @type {string}
   */
  title?: string;
  /**
   * Title text style.
   * @type {StyleProp<TextStyle>}
   */
  titleStyle?: StyleProp<TextStyle>;
  /**
   * Description text.
   * @type {string}
   */
  description?: string;
  /**
   * Description text style.
   * @type {StyleProp<TextStyle>}
   */
  descriptionStyle?: StyleProp<TextStyle>;
  /**
   * Background color.
   * @type {string}
   */
  backgroundColor: string;
  /**
   * Show close button when page active.
   * `note: last page will always show close button.`
   * @type {boolean}
   * @default false
   */
  showCloseButton?: boolean;
}

export interface PaperOnboardingScreenDimensions {
  width: number;
  height: number;
}
// ...
export interface PageProps {
  index: number;
  item: PaperOnboardingItemType;
  animatedIndex: SharedValue<number>;
  indicatorSize: number;
  titleStyle?: StyleProp<TextStyle>;
  descriptionStyle?: StyleProp<TextStyle>;
  screenDimensions: PaperOnboardingScreenDimensions;
  safeInsets: Required<Insets>;
}

export interface PageContentProps extends Pick<
  PaperOnboardingItemType,
  'title' | 'titleStyle' | 'description' | 'descriptionStyle' | 'image'
> {
  index: number;
  animatedFocus: SharedValue<number>;
}

export interface IndicatorIconProps {
  size: number;
}